CARTPOLE - utils gets optimizer, a2c_sync NN update after done
