+++ utils.py
@@ -49,6 +49,30 @@ def push_and_pull(opt, lnet, gnet, done, s_, bs, ba, br, gamma):
     lnet.load_state_dict(gnet.state_dict())
 
 
+def optimize(opt, lnet, done, s_, bs, ba, br, gamma):
+    if done:
+        v_s_ = 0.               # terminal
+    else:
+        v_s_ = lnet.forward(v_wrap(s_[None, :]))[-1].data.numpy()[0, 0]
+
+    buffer_v_target = []
+    for r in br[::-1]:    # reverse buffer r
+        # Advantage function?
+        v_s_ = r + gamma * v_s_
+        buffer_v_target.append(v_s_)
+    buffer_v_target.reverse()
+
+    loss = lnet.loss_func(
+        v_wrap(np.vstack(bs)),
+        v_wrap(np.array(ba), dtype=np.int64) if ba[0].dtype == np.int64 else v_wrap(np.vstack(ba)),
+        v_wrap(np.array(buffer_v_target)[:, None]))
+
+    # calculate local gradients
+    opt.zero_grad()
+    loss.backward()
+    opt.step()
+
+
 def record(global_ep, global_ep_r, ep_r, res_queue, name):
     with global_ep.get_lock():
         global_ep.value += 1
